worker/requirements.txt
argparse==1.2.1
pymongo==2.5
python-crontab==1.5.1
python-dateutil==2.2
six==1.4.1
wsgiref==0.1.2
